Switch over to npm/mill-bundler for js deps
Though it doesn't actually take advantage of the fancy newer linker in
mill, which would be our hope in terms of making things faster.
Looking at github.com/nafg/mill-bundler, it still uses fastOpt/fullOpt.
Probably if we wanted to move forward with this we could just contribute
some modifications, but it might be non-trivial (I spent hours trying
to figure out an apparent linking issue with js-cookie, to no avail).