Fix lobby tab jumping issues
Fixes #52.

Caused by re-rendering with initial state values when
receiving lobby updates from the server. Ideally I
should change it so re-rendering doesn't happen at all,
but this is a quick fix which has some useful code anyway
for syncing local state values with local storage.

There is still some annoying flicker on the leaderboard
page because it re-executes the async call. Would be
better to change that to websockets.

Also gets rid of the dependency on cookies, which I
don't need (somehow I'm still stuck in the 2000s...)